Move repo update scheduler into cmd/repo-updater
Only used in there, this makes it clearer (and more obvious what we still need to move where to make repo-updater do less).

## Test plan

Just moved code, CI should find any issues.
