+++ cmd/repo-updater/internal/scheduler/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
         "scheduler.go",
         "updatequeue.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/repos/scheduler",
-    visibility = ["//:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/repo-updater/internal/scheduler",
+    visibility = ["//cmd/repo-updater:__subpackages__"],
     deps = [
         "//internal/actor",
         "//internal/api",
