+++ cmd/repo-updater/shared/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
         "//cmd/repo-updater/internal/repoupdater",
+        "//cmd/repo-updater/internal/scheduler",
         "//internal/actor",
         "//internal/api",
         "//internal/authz",
@@ -42,7 +43,6 @@ go_library(
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/repos",
-        "//internal/repos/scheduler",
         "//internal/repoupdater/v1:repoupdater",
         "//internal/service",
         "//internal/trace",
